requirements.txt
asgiref==3.11.0
distlib==0.4.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
filelock==3.24.3
mysql-connector-python==9.5.0
packaging==26.0
platformdirs==4.9.2
PyMySQL==1.1.2
setuptools==80.10.2
sqlparse==0.5.5
tzdata==2025.3
virtualenv==20.38.0
virtualenvwrapper-win==1.2.7
wheel==0.46.3
